Refactored StudentService to be an interface
